requirements.txt
faiss-cpu==1.6.5
grpcio==1.35.0
grpcio-reflection==1.35.0
grpcio-health-checking==1.35.0
grpcio-tools==1.35.0
grpcio-channelz ==1.35.0
schema-entry==0.1.4
pyloggerhelper==0.0.2
numpy==1.19.5
uvloop==0.14.0
watchdog==2.0.2
readerwriterlock==1.0.8
Cython==0.29.22